requirements.txt
# Framework principal
fastapi==0.115.0
uvicorn[standard]==0.30.6

# MongoDB (driver async)
motor==3.6.0

# Validación y configuración
pydantic[email]==2.9.2
python-dotenv==1.0.1
pydantic-settings==2.10.1
python-jose==3.5.0
python-multipart==0.0.20

# Seguridad
PyJWT==2.9.0
passlib[bcrypt]==1.7.4

# Testing
pytest==8.3.3
httpx==0.27.2

# formateo/lint
black==24.8.0
flake8==7.1.1